overflow menu issue , total sale and collection issur in dashboard

+++ Benchmark/app/src/main/java/com/ae/benchmark/localdb/DBManager.java
@@ -198,7 +198,7 @@ public class DBManager {
         }
     }
 
-    public int getAllInvoiceHeadCollectionToday() {
+    public double getAllInvoiceHeadCollectionToday() {
 
         ArrayList<SalesInvoice> list = new ArrayList<>();
         openDatabsse();
@@ -220,10 +220,10 @@ public class DBManager {
             Log.e("dbError", e.toString());
         }
 
-        return (int) totColl;
+        return totColl;
     }
 
-    public int getAllInvoiceHeadCollection() {
+    public double getAllInvoiceHeadCollection() {
 
         ArrayList<SalesInvoice> list = new ArrayList<>();
         openDatabsse();
@@ -245,7 +245,7 @@ public class DBManager {
             Log.e("dbError", e.toString());
         }
 
-        return (int) totColl;
+        return totColl;
     }
     //INSERT SALESMAN
     public void insertSalesInvoiceHeader(SalesInvoice salesInvoice) {
